Changed the port of animation.json and the path of the login page
Had a problem with the project, because it did not load. I thought at first that it was from the servers, then from the broswers, then from the paths of the json files, but then found out that the problem was in the routing- it made rendering loops in the App component, which lead to it not rendering. But i fixed it.