Added a new json- server, that streams on port 3005 and a new component Process
Added a new json file process.json and a component Process.jsx, that i put in the App.jsx as a path and in the Navbar.jsx as a link so it could be reachable with routing. For Rest API i used simple json server. Taking the information from process.json to be visuallized on the page.